fix: dtype issue

+++ src/scanpy/get/_kernels.py
@@ -15,7 +15,7 @@ def agg_sum_csr(
     indicator: CSRBase,
     data: CSRBase,
 ):
-    out = np.zeros((indicator.shape[0], data.shape[1]))
+    out = np.zeros((indicator.shape[0], data.shape[1]), dtype="float64")
     for cat_num in numba.prange(indicator.shape[0]):
         start_cat_idx = indicator.indptr[cat_num]
         stop_cat_idx = indicator.indptr[cat_num + 1]
@@ -27,7 +27,7 @@ def agg_sum_csr(
 
             for j in range(start_obs, end_obs):
                 col = data.indices[j]
-                out[cat_num, col] += data.data[j]
+                out[cat_num, col] += float(data.data[j])
     return out
 
 
@@ -36,7 +36,7 @@ def agg_sum_csc(
     indicator: CSRBase,
     data: CSCBase,
 ):
-    out = np.zeros((indicator.shape[0], data.shape[1]), dtype=data.data.dtype)
+    out = np.zeros((indicator.shape[0], data.shape[1]), dtype="float64")
 
     obs_to_cat = np.full(data.shape[0], -1, dtype=np.int64)
 
@@ -53,6 +53,6 @@ def agg_sum_csc(
             cat = obs_to_cat[obs]
 
             if cat != -1:
-                out[cat, col] += data.data[j]
+                out[cat, col] += float(data.data[j])
 
     return out
